move memo from core to compat
