+++ compat/index.ts
@@ -1,4 +1,5 @@
 import { lazy, Suspense } from './suspense'
 import { createContext, useContext } from './context'
+import { memo } from './memo'
 
-export { lazy, Suspense, createContext, useContext }
+export { lazy, Suspense, createContext, useContext, memo }
